pyproject.toml
[tool.poetry]
name = "python-interview"
version = "0.1.0"
description = ""
authors = ["Michael Gagliardo <mgagliardo91@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
Faker = "^4.1.2"
pendulum = "^2.1.2"
black = "^19.10b0"
isort = "^5.4.2"
flake8 = "^3.8.3"
mypy = "^0.782"
poethepoet = "^0.7.0"
pre-commit = "^2.6.0"
pyyaml = "^5.3.1"

[tool.poe.tasks]
run = { script = "app.__main__:run", help = "Run main" }
lint = { cmd = "pre-commit run --all-files", help = "Run linters and formatters" }
test = { cmd = "pytest", help = "Run tests" }

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
